+++ client/src/routes/_pathlessLayout/posts/new.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from "@tanstack/react-router";
+import Form from "../../../component/form";
+
+export const Route = createFileRoute("/_pathlessLayout/posts/new")({
+  component: Form,
+});
